add rating basic logic
